requirements.txt
streamlit>=1.33
psycopg[binary]>=3.1
openai>=1.40
pandas>=2.2
python-dotenv>=1.0